Use mediump instead of highp See https://www.khronos.org/registry/OpenGL/specs/es/3.0/GLSL_ES_Specification_3.00.pdf#page=60

+++ src/webgl/texture/combine-bands.js
@@ -27,9 +27,9 @@ uniform sampler2D bitmapTexture_a;
 `;
 
 const fs2 = `\
-precision highp float;
-precision highp int;
-precision highp usampler2D;
+precision mediump float;
+precision mediump int;
+precision mediump usampler2D;
 
 uniform usampler2D bitmapTexture_r;
 uniform usampler2D bitmapTexture_g;
